Reorganize and streamline
* static BGP session methods moved into BgpSessionInfo
* abstract BgpSessionQuestion class with methods both BGP session
questions share
