+++ projects/question/src/main/java/org/batfish/question/bgpsessionstatus/BgpSessionCompatibilityAnswerer.java
@@ -1,6 +1,7 @@
 package org.batfish.question.bgpsessionstatus;
 
-import com.google.common.annotations.VisibleForTesting;
+import static org.batfish.question.bgpsessionstatus.BgpSessionInfo.getBgpSessionInfo;
+
 import com.google.common.collect.HashMultiset;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Multiset;
@@ -12,24 +13,16 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import org.batfish.common.Answerer;
 import org.batfish.common.BatfishException;
 import org.batfish.common.plugin.IBatfish;
 import org.batfish.common.util.CommonUtil;
-import org.batfish.datamodel.BgpActivePeerConfig;
-import org.batfish.datamodel.BgpPassivePeerConfig;
-import org.batfish.datamodel.BgpPeerConfig;
 import org.batfish.datamodel.BgpPeerConfigId;
 import org.batfish.datamodel.BgpSessionProperties;
-import org.batfish.datamodel.BgpSessionProperties.SessionType;
 import org.batfish.datamodel.Configuration;
-import org.batfish.datamodel.Interface;
 import org.batfish.datamodel.Ip;
-import org.batfish.datamodel.NetworkConfigurations;
 import org.batfish.datamodel.answers.AnswerElement;
 import org.batfish.datamodel.answers.Schema;
-import org.batfish.datamodel.collections.NodeInterfacePair;
 import org.batfish.datamodel.pojo.Node;
 import org.batfish.datamodel.questions.DisplayHints;
 import org.batfish.datamodel.questions.Question;
@@ -38,7 +31,6 @@ import org.batfish.datamodel.table.Row;
 import org.batfish.datamodel.table.Row.RowBuilder;
 import org.batfish.datamodel.table.TableAnswerElement;
 import org.batfish.datamodel.table.TableMetadata;
-import org.batfish.question.bgpsessionstatus.BgpSessionInfo.SessionStatus;
 
 public class BgpSessionCompatibilityAnswerer extends Answerer {
 
@@ -117,125 +109,6 @@ public class BgpSessionCompatibilityAnswerer extends Answerer {
             .collect(ImmutableList.toImmutableList()));
 
     return sessions;
-  }
-
-  private static Optional<BgpSessionInfo> getBgpSessionInfo(
-      BgpSessionCompatibilityQuestion question,
-      Map<String, Configuration> configurations,
-      Set<String> includeNodes1,
-      Set<String> includeNodes2,
-      Map<Ip, Set<String>> ipOwners,
-      Set<Ip> allInterfaceIps,
-      ValueGraph<BgpPeerConfigId, BgpSessionProperties> configuredBgpTopology,
-      ValueGraph<BgpPeerConfigId, BgpSessionProperties> establishedBgpTopology,
-      BgpPeerConfigId bgpPeerConfigId) {
-    String hostname = bgpPeerConfigId.getHostname();
-    String vrfName = bgpPeerConfigId.getVrfName();
-    // Only match nodes we care about
-    if (!includeNodes1.contains(hostname)) {
-      return Optional.empty();
-    }
-
-    NetworkConfigurations networkConfigurations = NetworkConfigurations.of(configurations);
-    BgpPeerConfig bgpPeerConfig = networkConfigurations.getBgpPeerConfig(bgpPeerConfigId);
-    if (bgpPeerConfig == null) {
-      return Optional.empty();
-    }
-    // Only match groups we care about
-    if (bgpPeerConfig.getGroup() != null
-        && question.matchesForeignGroup(bgpPeerConfig.getGroup())) {
-      return Optional.empty();
-    }
-
-    // Setup session info.
-    SessionType sessionType =
-        bgpPeerConfig instanceof BgpActivePeerConfig
-            ? BgpSessionProperties.getSessionType((BgpActivePeerConfig) bgpPeerConfig)
-            : SessionType.UNSET;
-
-    // Skip session types we don't care about
-    if (!question.matchesType(sessionType)) {
-      return Optional.empty();
-    }
-
-    BgpSessionInfo.Builder bsiBuilder =
-        BgpSessionInfo.builder(
-            hostname, vrfName, bgpPeerConfigId.getRemotePeerPrefix(), sessionType);
-
-    SessionStatus configuredStatus;
-    if (bgpPeerConfig instanceof BgpPassivePeerConfig) {
-      configuredStatus = SessionStatus.DYNAMIC_LISTEN;
-    } else if (bgpPeerConfig instanceof BgpActivePeerConfig) {
-      configuredStatus = getLocallyBrokenStatus((BgpActivePeerConfig) bgpPeerConfig, sessionType);
-    } else {
-      throw new BatfishException("Unsupported type of BGP peer config (not active or passive)");
-    }
-
-    if (configuredStatus == null) {
-      /*
-       * Nothing blatantly broken so far on the local side, keep checking.
-       * Also at this point we know this is not a Dynamic bgp neighbor
-       */
-      Ip localIp = bgpPeerConfig.getLocalIp();
-      bsiBuilder.withLocalIp(localIp);
-      Optional<Interface> iface =
-          CommonUtil.getActiveInterfaceWithIp(localIp, configurations.get(hostname));
-      bsiBuilder.withLocalInterface(
-          iface
-              .map(anInterface -> new NodeInterfacePair(hostname, anInterface.getName()))
-              .orElse(null));
-
-      BgpActivePeerConfig p2pBgpPeerConfig = (BgpActivePeerConfig) bgpPeerConfig;
-      Ip remoteIp = p2pBgpPeerConfig.getPeerAddress();
-
-      if (!allInterfaceIps.contains(localIp)) {
-        configuredStatus = SessionStatus.INVALID_LOCAL_IP;
-      } else if (remoteIp == null || !allInterfaceIps.contains(remoteIp)) {
-        configuredStatus = SessionStatus.UNKNOWN_REMOTE;
-      } else {
-        if (!node2RegexMatchesIp(remoteIp, ipOwners, includeNodes2)) {
-          return Optional.empty();
-        }
-        if (configuredBgpTopology.adjacentNodes(bgpPeerConfigId).isEmpty()) {
-          configuredStatus = SessionStatus.HALF_OPEN;
-          // degree > 2 because of directed edges. 1 edge in, 1 edge out == single connection
-        } else if (configuredBgpTopology.degree(bgpPeerConfigId) > 2) {
-          configuredStatus = SessionStatus.MULTIPLE_REMOTES;
-        } else {
-          BgpPeerConfigId remoteNeighbor =
-              configuredBgpTopology.adjacentNodes(bgpPeerConfigId).iterator().next();
-          bsiBuilder.withRemoteNode(remoteNeighbor.getHostname());
-          configuredStatus = SessionStatus.UNIQUE_MATCH;
-        }
-      }
-    }
-    if (!question.matchesStatus(configuredStatus)) {
-      return Optional.empty();
-    }
-
-    bsiBuilder.withConfiguredStatus(configuredStatus);
-
-    bsiBuilder.withEstablishedNeighbors(
-        establishedBgpTopology != null && establishedBgpTopology.nodes().contains(bgpPeerConfigId)
-            ? establishedBgpTopology.inDegree(bgpPeerConfigId)
-            : -1);
-    return Optional.of(bsiBuilder.build());
-  }
-
-  @Nullable
-  @VisibleForTesting
-  static SessionStatus getLocallyBrokenStatus(
-      BgpActivePeerConfig neighbor, SessionType sessionType) {
-    if (neighbor.getLocalIp() == null) {
-      if (sessionType == SessionType.EBGP_MULTIHOP || sessionType == SessionType.IBGP) {
-        return SessionStatus.LOCAL_IP_UNKNOWN_STATICALLY;
-      } else {
-        return SessionStatus.NO_LOCAL_IP;
-      }
-    } else if (neighbor.getRemoteAs() == null) {
-      return SessionStatus.NO_REMOTE_AS;
-    }
-    return null;
   }
 
   public static TableMetadata createMetadata(Question question) {
