+++ projects/question/src/main/java/org/batfish/question/bgpsessionstatus/BgpSessionInfo.java
@@ -4,12 +4,28 @@ import static java.util.Objects.requireNonNull;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.google.common.annotations.VisibleForTesting;
+import com.google.common.collect.Sets;
+import com.google.common.graph.ValueGraph;
 import java.util.Comparator;
+import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import org.batfish.common.BatfishException;
+import org.batfish.common.util.CommonUtil;
+import org.batfish.datamodel.BgpActivePeerConfig;
+import org.batfish.datamodel.BgpPassivePeerConfig;
+import org.batfish.datamodel.BgpPeerConfig;
+import org.batfish.datamodel.BgpPeerConfigId;
+import org.batfish.datamodel.BgpSessionProperties;
 import org.batfish.datamodel.BgpSessionProperties.SessionType;
+import org.batfish.datamodel.Configuration;
+import org.batfish.datamodel.Interface;
 import org.batfish.datamodel.Ip;
+import org.batfish.datamodel.NetworkConfigurations;
 import org.batfish.datamodel.Prefix;
 import org.batfish.datamodel.collections.NodeInterfacePair;
 
@@ -78,6 +94,135 @@ public class BgpSessionInfo implements Comparable<BgpSessionInfo> {
         requireNonNull(remotePrefix, PROP_REMOTE_PREFIX + " must not be null"),
         requireNonNull(sessionType, PROP_SESSION_TYPE + " must not be null"),
         requireNonNull(vrfName, PROP_VRF_NAME + " must not be null"));
+  }
+
+  static Optional<BgpSessionInfo> getBgpSessionInfo(
+      BgpSessionQuestion question,
+      Map<String, Configuration> configurations,
+      Set<String> includeNodes1,
+      Set<String> includeNodes2,
+      Map<Ip, Set<String>> ipOwners,
+      Set<Ip> allInterfaceIps,
+      ValueGraph<BgpPeerConfigId, BgpSessionProperties> configuredBgpTopology,
+      ValueGraph<BgpPeerConfigId, BgpSessionProperties> establishedBgpTopology,
+      BgpPeerConfigId bgpPeerConfigId) {
+    String hostname = bgpPeerConfigId.getHostname();
+    String vrfName = bgpPeerConfigId.getVrfName();
+    // Only match nodes we care about
+    if (!includeNodes1.contains(hostname)) {
+      return Optional.empty();
+    }
+
+    NetworkConfigurations networkConfigurations = NetworkConfigurations.of(configurations);
+    BgpPeerConfig bgpPeerConfig = networkConfigurations.getBgpPeerConfig(bgpPeerConfigId);
+    if (bgpPeerConfig == null) {
+      return Optional.empty();
+    }
+    // Only match groups we care about
+    if (bgpPeerConfig.getGroup() != null
+        && question.matchesForeignGroup(bgpPeerConfig.getGroup())) {
+      return Optional.empty();
+    }
+
+    // Setup session info.
+    SessionType sessionType =
+        bgpPeerConfig instanceof BgpActivePeerConfig
+            ? BgpSessionProperties.getSessionType((BgpActivePeerConfig) bgpPeerConfig)
+            : SessionType.UNSET;
+
+    // Skip session types we don't care about
+    if (!question.matchesType(sessionType)) {
+      return Optional.empty();
+    }
+
+    BgpSessionInfo.Builder bsiBuilder =
+        BgpSessionInfo.builder(
+            hostname, vrfName, bgpPeerConfigId.getRemotePeerPrefix(), sessionType);
+
+    SessionStatus configuredStatus;
+    if (bgpPeerConfig instanceof BgpPassivePeerConfig) {
+      configuredStatus = SessionStatus.DYNAMIC_LISTEN;
+    } else if (bgpPeerConfig instanceof BgpActivePeerConfig) {
+      configuredStatus = getLocallyBrokenStatus((BgpActivePeerConfig) bgpPeerConfig, sessionType);
+    } else {
+      throw new BatfishException("Unsupported type of BGP peer config (not active or passive)");
+    }
+
+    if (configuredStatus == null) {
+      /*
+       * Nothing blatantly broken so far on the local side, keep checking.
+       * Also at this point we know this is not a Dynamic bgp neighbor
+       */
+      Ip localIp = bgpPeerConfig.getLocalIp();
+      bsiBuilder.withLocalIp(localIp);
+      Optional<Interface> iface =
+          CommonUtil.getActiveInterfaceWithIp(localIp, configurations.get(hostname));
+      bsiBuilder.withLocalInterface(
+          iface
+              .map(anInterface -> new NodeInterfacePair(hostname, anInterface.getName()))
+              .orElse(null));
+
+      BgpActivePeerConfig p2pBgpPeerConfig = (BgpActivePeerConfig) bgpPeerConfig;
+      Ip remoteIp = p2pBgpPeerConfig.getPeerAddress();
+
+      if (!allInterfaceIps.contains(localIp)) {
+        configuredStatus = SessionStatus.INVALID_LOCAL_IP;
+      } else if (remoteIp == null || !allInterfaceIps.contains(remoteIp)) {
+        configuredStatus = SessionStatus.UNKNOWN_REMOTE;
+      } else {
+        if (!node2RegexMatchesIp(remoteIp, ipOwners, includeNodes2)) {
+          return Optional.empty();
+        }
+        if (configuredBgpTopology.adjacentNodes(bgpPeerConfigId).isEmpty()) {
+          configuredStatus = SessionStatus.HALF_OPEN;
+          // degree > 2 because of directed edges. 1 edge in, 1 edge out == single connection
+        } else if (configuredBgpTopology.degree(bgpPeerConfigId) > 2) {
+          configuredStatus = SessionStatus.MULTIPLE_REMOTES;
+        } else {
+          BgpPeerConfigId remoteNeighbor =
+              configuredBgpTopology.adjacentNodes(bgpPeerConfigId).iterator().next();
+          bsiBuilder.withRemoteNode(remoteNeighbor.getHostname());
+          configuredStatus = SessionStatus.UNIQUE_MATCH;
+        }
+      }
+    }
+    if (!question.matchesStatus(configuredStatus)) {
+      return Optional.empty();
+    }
+
+    bsiBuilder.withConfiguredStatus(configuredStatus);
+
+    bsiBuilder.withEstablishedNeighbors(
+        establishedBgpTopology != null && establishedBgpTopology.nodes().contains(bgpPeerConfigId)
+            ? establishedBgpTopology.inDegree(bgpPeerConfigId)
+            : -1);
+    return Optional.of(bsiBuilder.build());
+  }
+
+  @Nullable
+  @VisibleForTesting
+  static SessionStatus getLocallyBrokenStatus(
+      BgpActivePeerConfig neighbor, SessionType sessionType) {
+    if (neighbor.getLocalIp() == null) {
+      if (sessionType == BgpSessionProperties.SessionType.EBGP_MULTIHOP
+          || sessionType == BgpSessionProperties.SessionType.IBGP) {
+        return SessionStatus.LOCAL_IP_UNKNOWN_STATICALLY;
+      } else {
+        return SessionStatus.NO_LOCAL_IP;
+      }
+    } else if (neighbor.getRemoteAs() == null) {
+      return SessionStatus.NO_REMOTE_AS;
+    }
+    return null;
+  }
+
+  private static boolean node2RegexMatchesIp(
+      Ip ip, Map<Ip, Set<String>> ipOwners, Set<String> includeNodes2) {
+    Set<String> owners = ipOwners.get(ip);
+    if (owners == null) {
+      throw new BatfishException("Expected at least one owner of ip: " + ip);
+    }
+    return !Sets.intersection(includeNodes2, owners).isEmpty();
   }
 
   private BgpSessionInfo(
